Analysis/requirements.txt
google-ads
pandas
python-dotenv
